requirements.txt
azure_storage==0.37.0
python-telegram-bot==20.2
Requests==2.29.0
